+++ include/vmpl.h
@@ -170,9 +170,9 @@ extern void dune_setcontext(const ucontext_t *ucp, struct dune_tf *tf);
 extern bool register_syscall_filter(bool (*filter)(struct dune_tf *tf));
 
 // hotcall routines
-void register_hotcall(int syscall);
-void unregister_hotcall(int syscall);
-bool is_hotcall(int syscall);
+void register_hotcall(long syscall);
+void unregister_hotcall(long syscall);
+bool is_hotcall(long syscall);
 long vmpl_hotcalls_call(struct dune_tf *tf);
 
 // elf helper functions
